Move classes to composite
