+++ jdi-light-bootstrap/src/main/java/com/epam/jdi/light/ui/bootstrap/asserts/BootstrapDropdownAssert.java
@@ -2,11 +2,7 @@ package com.epam.jdi.light.ui.bootstrap.asserts;
 
 import com.epam.jdi.light.asserts.generic.UIAssert;
 import com.epam.jdi.light.common.JDIAction;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.BootstrapDropdown;
-import org.hamcrest.Matcher;
-import org.hamcrest.Matchers;
-
-import java.util.Arrays;
+import com.epam.jdi.light.ui.bootstrap.elements.composite.BootstrapDropdown;
 
 import static com.epam.jdi.light.asserts.core.SoftAssert.jdiAssert;
 import static org.hamcrest.Matchers.is;
